complete init_redis
